fix CCColor componentRGBA crashing on 64 bit

+++ Test/SpriteKitReader-Experimental/SpriteKitReader-Experimental/CCBReader/SpriteKit/CCColor.h
@@ -109,7 +109,7 @@
 @end
 
 typedef struct {
-    CGFloat r, g, b, a;
+    float r, g, b, a;
 } CCColorComponentRGBA;
 
 @interface CCColor (CCBReader)
